LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-activation/activation-traverse-then-clobber-expected.txt



FAIL navigation.activation - traverse, then push same-document assert_equals: expected 2 but got 1
